Refactor githubRaw
There are at least three levels of information that we need:

1) Project board with numbers of issues in each column.
2) Project board with all issues.
3) Project board with all issues, additionally with all reactions.

There are now different types to represent all of these, so that data
can be passed around in a type-safe manner. It's *not* safe to have
something like

type issue = { ..., reactions : (string * person) list option }

because if we need reactions and issue.reactions is None, then we get
a runtime error. So we need to have two types:

type issue = { ... }
type issue_with_reactions = { ..., reactions : (string * person) list }

This commit implements all the necessary types for the different levels
of data required.