Attempt to pick an appropriate finance code from Forecast
If more than one tag is present on Forecast, it's not obvious which one
to chose as the finance code. This commit implements a rough heuristic
to pick out the 'correct' finance code (and issue a debug message when
it does this). It has a 100% success rate on existing projects, but a
warning about having multiple finance codes is still issued anyway just
to be safe.

Closes #78