Change CLI
You now need `--notify github --all-github` in order to actually post to
all GitHub issues. Just doing `--notify github` will only work on the
four blessed columns.

This should probably be reworked, to be honest. We should have the
`issues` argument take care of all of this:

 --issues all
 --issues columns (default)
 --issues NUM1,NUM2,...

To do this, we'd need a custom parser, to be written tomorrow.