Rework get_all_users_async
Fixes #99

The function now uses the REST API and also makes sure to fetch all
possible users, not just the first 100.

The justification for using GraphQL was that it was faster than the REST
API. However, this is only definitely true if there are fewer than 100
users which need to be fetched. Because GraphQL cannot return more than
100 users at a go, obtaining all users would require at least two HTTP
requests, plus some extra, very hacky, work to manage the pagination. It
is probably easier to just use the REST API.